(test) moved route tests to integration.js, changed server.js to dbSpec.js
